% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cme.R
\name{cmeCycle2}
\alias{cmeCycle2}
\title{cmeCycle2}
\usage{
cmeCycle2(
  ylist,
  xlist,
  zlist,
  beta,
  errVar,
  a0,
  b0,
  lambda2,
  delta2,
  nu,
  xi,
  gammaSamp,
  R,
  S
)
}
\arguments{
\item{ylist}{a list of response vectors where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{xlist}{a list of fixed effect covariates where the \eqn{i^{th}} component contains \eqn{m_i \times p} design matrix for the \eqn{i^{th}} subject.}

\item{zlist}{a list of random effect covariates where the \eqn{i^{th}} component contains \eqn{m_i \times q} design matrix for the \eqn{i^{th}} subject.}

\item{beta}{a \eqn{p}-dimensional vector for beta value from previous iteration.}

\item{errVar}{error variance value from previous iteration.}

\item{a0}{shape hyperparameter for inverse-gamma prior for error variance.}

\item{b0}{scale hyperparameter for inverse-gamma prior for error variance.}

\item{lambda2}{\eqn{p}-dimensional vector for squared local shrinkage parameter values from previous iteration.}

\item{delta2}{global shrinkage parameter value from previous iteration.}

\item{nu}{a \eqn{p}-dimensional vector for auxiliary variable \eqn{\nu_{j}}'s from previous iteration.}

\item{xi}{value of auxiliary variable \eqn{\xi} from previous iteration.}

\item{gammaSamp}{a \eqn{k_1 k_2}-dimensional vector for updated posterior draw of columnwise vectorized \eqn{\Gamma} from first cycle.}

\item{R}{\eqn{k_2 \times q} random projection matrix with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_2}.}

\item{S}{\eqn{k_1 \times q} random projection matrix with entries iid from a Gaussian distribution with mean 0 and variance \eqn{1/k_1}.}
}
\description{
This function implements the second cycle of the collapsed Gibbs sampler for the CME model by sampling the high-dimensional parameter \eqn{(\beta, \tau^2)} from their full-conditional distributions.
}
