% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cometCycle1.R
\name{cometCycle1}
\alias{cometCycle1}
\title{cometCycle1}
\usage{
cometCycle1(
  y,
  vecxlist,
  comp_zlist,
  z_tilde_list,
  mis,
  B,
  errVar,
  Gamma_list,
  Sigma_gamma_list,
  kdims,
  RRtkron
)
}
\arguments{
\item{y}{a vector containing the response for all the observations.}

\item{vecxlist}{a list, each component contains the vectorized fixed-effect covariate for each observation.}

\item{comp_zlist}{a list, each component contains the compressed random-effect covariate tensor \eqn{\tilde{\mathcal{Z}}} for each observation.}

\item{z_tilde_list}{a list, each component contains vectorized compressed random-effect covariates, stacked for each cluster.}

\item{mis}{a vector of cluster sizes.}

\item{B}{\eqn{\mathcal{B}} sampled in previous iteration.}

\item{errVar}{\eqn{\tau^2} sampled in previous iteration.}

\item{Gamma_list}{a list of length \eqn{D}, each component contains \eqn{\Gamma_d} sampled in previous iteration.}

\item{Sigma_gamma_list}{a list of length \eqn{D}, each component corresponds to the prior covariance matrix for the corresponding \eqn{\gamma_d = \operatorname{vec}(\Gamma_d)}.}

\item{kdims}{vector of compressed covariance dimensions.}

\item{RRtkron}{\eqn{R_D R_D^{\top} \otimes \dots \otimes R_1 R_1^{\top}} computed at the start of the model fitting.}
}
\value{
a list containing the following components:
\describe{
\item{GammaSamplist}{a list of \eqn{D} matrices each of dimension \eqn{k_d \times k_d}, denoting the updated compressed covariance parameters at a given iteration.}
\item{Di_tilde}{an array of dimension \eqn{k_1 \times \dots \times k_D}, denoting the imputed compressed random slope tensor at a given iteration. }
}
}
\description{
This function implements the collapsed Gibbs sampler for fitting the compressed mixed-effects tensor (CoMET) model.
}
