% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cometCycle2.R
\name{cometCycle2}
\alias{cometCycle2}
\title{cometCycle2}
\usage{
cometCycle2(
  yijs,
  xijlist,
  zi_tilde_list,
  mis,
  B_factors,
  K,
  a0,
  b0,
  errVar,
  lambda2_list,
  delta2_list,
  nu_list,
  xi_list,
  Gkron,
  RRtkron
)
}
\arguments{
\item{yijs}{a vector containing the response for all the observations.}

\item{xijlist}{a list, each component contains a tensor-valued fixed-effect covariate for each observation.}

\item{zi_tilde_list}{a list, each component contains vectorized compressed random-effect covariates, stacked for each cluster.}

\item{mis}{a vector of cluster sizes.}

\item{B_factors}{fixed-effect factor matrices sampled in previous iteration.}

\item{K}{a user-specified CP rank of fixed-effect coefficient \eqn{\mathcal{B}}. Should not exceed the dimensions for matrix-valued covariates.}

\item{a0}{shape hyperparameter for inverse-gamma prior for idiosyncratic error variance.}

\item{b0}{scale hyperparameter for inverse-gamma prior for idiosyncratic error variance.}

\item{errVar}{idiosyncratic error variance sampled in previous iteration.}

\item{lambda2_list}{a list of length \eqn{D}, each containing \eqn{K p_d} squared local shrinkage \eqn{\lambda^2} values from previous iteration.}

\item{delta2_list}{a vector of length \eqn{K}, value of rank-\eqn{g}-specific global shrinkage \eqn{\delta^2} values from previous iteration.}

\item{nu_list}{a list of length \eqn{D}, each containing \eqn{K p_d} auxiliary variable \eqn{\nu} values from previous iteration.}

\item{xi_list}{a vector of length \eqn{K}, value of rank-\eqn{g}-specific auxiliary variable \eqn{\xi_g} values from previous iteration.}

\item{Gkron}{\eqn{\Gamma_D \Gamma_D^{\top} \otimes \dots \otimes \Gamma_1 \Gamma_1^{\top}} computed using the \eqn{\Gamma_d} updates from the previous iteration.}

\item{RRtkron}{\eqn{R_D R_D^{\top} \otimes \dots \otimes R_1 R_1^{\top}} computed at the start of the model fitting.}
}
\value{
a list containing the following components:
\describe{
\item{BfactorsSamp}{a list of length \eqn{D}, each component comprising the posterior draw of \eqn{p_d \times K} factor matrix \eqn{B_d}.}
\item{BSamp}{an array denoting the updated fixed-effect parameter \eqn{\mathcal{B}}.}
\item{errVarSamp}{the updated posterior draw of error variance \eqn{\tau^2}.}
\item{lambda2Samplist}{a list of length \eqn{D}, each component containing the updated \eqn{K p_d} squared local shrinkage \eqn{\lambda^2} draws.}
\item{delta2Samplist}{a vector of length \eqn{K} storing the updated rank-\eqn{g}-specific global shrinkage \eqn{\delta^2} draws.}
\item{nuSamplist}{a list of length \eqn{D}, each component containing the updated \eqn{K p_d} auxiliary variable \eqn{nu} draws.}
\item{xiSamplist}{a vector of length \eqn{K} storing the updated rank-\eqn{g}-specific auxiliary variable \eqn{\xi_g} draws.}
}
}
\description{
This function implements the collapsed Gibbs sampler for fitting the compressed mixed-effects tensor (CoMET) model.
}
