% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cpB}
\alias{cpB}
\title{cpB}
\usage{
cpB(factorsList)
}
\arguments{
\item{factorsList}{list of factor matrices.}
}
\value{
a tensor
}
\description{
Given a list of factor matrices, each of dimension \eqn{p_d \times K}, reconstructs a \eqn{p_1 \times \dots \times p_D} tensor \eqn{\mathcal{B}} using rank-\eqn{K} CP/PARAFAC structure.
}
