% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{khatri_rao_comet}
\alias{khatri_rao_comet}
\title{khatri_rao_comet}
\usage{
khatri_rao_comet(A, B)
}
\arguments{
\item{A}{a \eqn{p_1 \times k} matrix.}

\item{B}{a \eqn{p_2 \times k} matrix.}
}
\value{
a \eqn{p_1 p_2 \times k} matrix with the \eqn{d}-th column as Kronecker product of the \eqn{d}-th columns of \eqn{A} and \eqn{B}.
}
\description{
matching-columnwise Khatri-Rao product of two matrices with same number of columns.
}
