% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_wrappers.R
\name{LOOrevkron}
\alias{LOOrevkron}
\title{LOOrevkron}
\usage{
LOOrevkron(Glist, d)
}
\arguments{
\item{Glist}{a list of matrices}

\item{d}{an integer denoting the mode to be left out.}
}
\value{
a matrix yielding kronecker product in reverse order.
}
\description{
Returns \eqn{G_n \otimes G_{d+1} \otimes G_{d-1} \dots G_1}
}
