% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mode_matricize}
\alias{mode_matricize}
\title{mode_matricize}
\usage{
mode_matricize(A, d)
}
\arguments{
\item{A}{a numeric array with \code{dim} attribute.}

\item{d}{integer, the mode index to be matricized.}
}
\value{
a matrix corresponding to the mode-\eqn{d} unfolding.
}
\description{
mode-n matricization of a tensor
}
