% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_existingsubj.R
\name{predict_existingsubj}
\alias{predict_existingsubj}
\title{predict_existingsubj}
\usage{
predict_existingsubj(
  object,
  kdims,
  R_list,
  S_list,
  xlist_test,
  zlist_test,
  mis,
  nom.level
)
}
\arguments{
\item{object}{a list of posterior samples of parameters returned by the function \code{\link{comet}}.}

\item{kdims}{a vector of length \eqn{D} (number of tensor modes), where each element equals the \eqn{d}-th mode-specific compressed covariance dimension \eqn{k_d}.}

\item{R_list}{the list of \eqn{k_d \times q} random projection matrices used for compressing the random slope in the Gibbs sampler of CoMET. The list should be same as that used as argument to the function \code{\link{comet}}.}

\item{S_list}{the list of \eqn{k_d \times q} random projection matrices used for compressing the random-effect covariates in the Gibbs sampler of CoMET. The list should be same as that used as argument to the function \code{\link{comet}}.}

\item{xlist_test}{a list of fixed effect covariates where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{zlist_test}{a list of random effect covariates where the \eqn{i^{th}} component contains \eqn{m_i} observations for the \eqn{i^{th}} subject.}

\item{mis}{a vector of cluster sizes of the new subjects.}

\item{nom.level}{nominal level for constructing prediction intervals.}
}
\value{
a list containing the following components:
\describe{
\item{ypred}{predicted values of response.}
\item{yhat_samples}{posterior samples of the response.}
\item{lower_pi}{lower limits of the prediction intervals.}
\item{upper_pi}{upper limits of the prediction intervals.}
}
}
\description{
This function performs posterior prediction of new observations for existing subjects using the compressed mixed model.
}
