% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{revkronAll}
\alias{revkronAll}
\title{revkronAll}
\usage{
revkronAll(Glist)
}
\arguments{
\item{Glist}{a list of matrices.}
}
\value{
a matrix yielding Kronecker product of in reverse order.
}
\description{
Given \eqn{n} matrices \eqn{G_1, \dots, G_n}, computes \eqn{G_n \otimes \dots \otimes G_1}.
}
