% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{revkronLOO}
\alias{revkronLOO}
\title{revkronLOO}
\usage{
revkronLOO(Glist, d)
}
\arguments{
\item{Glist}{a list of matrices.}

\item{d}{integer, index corresponding to the matrix to be excluded from the Kronecker product.}
}
\value{
a matrix yielding Kronecker product in reverse order, leaving the \eqn{d}-th matrix.
}
\description{
Given \eqn{n} matrices \eqn{G_1, \dots, G_n}, computes \eqn{G_n \otimes \dots \otimes G_{d+1} \otimes G_{d-1} \otimes \dots \otimes G_1}.
}
