% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comet.R
\name{comet}
\alias{comet}
\title{comet}
\usage{
comet(
  y,
  xlist,
  zlist,
  mis,
  K,
  kdims,
  a0,
  b0,
  gammaVar0,
  R_list,
  S_list,
  niter,
  nburn,
  nthin,
  store_ranef = FALSE
)
}
\arguments{
\item{y}{a vector containing the response for all the observations.}

\item{xlist}{a list, each component contains a tensor-valued fixed-effect covariate for each observation.}

\item{zlist}{a list, each component contains a tensor-valued random-effect covariate for each observation.}

\item{mis}{a vector of cluster sizes.}

\item{K}{a user-specified rank for CP structure of fixed-effect coefficient. Should not exceed the dimensions for matrix-valued covariates.}

\item{kdims}{a vector of length \eqn{D} (number of tensor modes), where each element equals the \eqn{d}-th mode-specific compressed covariance dimension \eqn{k_d}.}

\item{a0}{shape hyperparameter for inverse-gamma prior for idiosyncratic error variance.}

\item{b0}{scale hyperparameter for inverse-gamma prior for idiosyncratic error variance.}

\item{gammaVar0}{a vector of length \eqn{D} (number of tensor modes) with each element being the variance hyperparameter for the Gaussian prior on the columnwise vectorization of the mode-specific compressed covariance parameter.}

\item{R_list}{a list of length \eqn{D} (number of tensor modes). Each component corresponds to a \eqn{k_d \times q} random projection matrix (for compressing the random slope) with iid entries following \eqn{\mathcal{N}(0, 1/k_d)}, where \eqn{k_d} is the compressed covariance dimension for mode-\eqn{d}.}

\item{S_list}{a list of length \eqn{D} (number of tensor modes). Each component corresponds to a \eqn{k_d \times q} random projection matrix (for compressing the random-effect covariate) with iid entries following \eqn{\mathcal{N}(0, 1/k_d)}, where \eqn{k_d} is the compressed covariance dimension for mode-\eqn{d}.}

\item{niter}{number of MCMC iterations.}

\item{nburn}{number of burn-in samples.}

\item{nthin}{thinning size specifying every nthin-th sample is retained.}

\item{store_ranef}{logical; if TRUE, the imputed compressed random-effects are stored. Default value is FALSE.}
}
\value{
a list containing the following components:
\describe{
\item{betaSamp}{a matrix with \eqn{\bigl(\text{niter}-\text{nburn}\bigr)/\text{nthin}} rows and \eqn{p} columns containing posterior samples of fixed effect parameter \eqn{\beta}.}
\item{errVarSamp}{a numeric vector of length \eqn{\bigl(\text{niter}-\text{nburn}\bigr)/\text{nthin}} containing posterior samples of error variance \eqn{\tau^2}.}
\item{gammaSamp}{a matrix with \eqn{\bigl(\text{niter}-\text{nburn}\bigr)/\text{nthin}} rows and \eqn{k_1 k_2} columns containing posterior samples of columnwise vectorization of the compressed covariance parameter \eqn{\Gamma}.}
\item{ranefSamplist}{a list of length \eqn{\bigl(\text{niter}-\text{nburn}\bigr)/\text{nthin}}, containing the imputed vectorized compressed random-effects. Returned only when store_ranef = TRUE.}
\item{sampler_time}{time taken by the sampler to run.}
}
}
\description{
This function implements the collapsed Gibbs sampler for fitting the compressed mixed-effects tensor (CoMET) model.
}
