% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_wrappers.R
\name{khatri_rao_comet}
\alias{khatri_rao_comet}
\title{khatri_rao_comet}
\usage{
khatri_rao_comet(A, B)
}
\arguments{
\item{A}{a matrix of dimensions \eqn{p_1 \times k}.}

\item{B}{a matrix of dimensions \eqn{p_2 \times k}.}
}
\value{
a matrix of dimensions \eqn{p_1 p_2 \times k}.
}
\description{
matching-columnwise Khatri-Rao product
}
