% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_wrappers.R
\name{revkronAll}
\alias{revkronAll}
\title{revkronAll}
\usage{
revkronAll(Glist)
}
\arguments{
\item{Glist}{a list of matrices}
}
\value{
a matrix yielding kronecker product in reverse order.
}
\description{
Returns \eqn{G_n \otimes \dots G_1} for a list of n matrices.
}
